requirements.txt
camelot-py
opencv-python
beautifulsoup4
PyPDF2
sentence_transformers
langchain-community
langchain-huggingface